+++ newclient/src/modules/charts/handlers/TimeSeriesChartData.vue
@@ -6,6 +6,7 @@
             :datacollection="chartData"
             :chart-options="chartOptions"
             :is-pending="isPending"
+            :is-loading-data="loading"
             @timeFrame="setTimeFrame"
         />
     </div>
@@ -24,10 +25,11 @@ import {
     TimeseriesValue,
     TimeseriesValueKey,
     TimeOptions,
-    ComponentDataInterface
+    ComponentDataInterface,
+    ChartRouteKey
 } from '@app/modules/charts/models'
 import { TimeseriesScale } from '@app/apollo/global/globalTypes'
-import { getTimeseriesData_getTimeseriesData as GetTimeseriesDataType } from '@app/modules/charts/handlers/apolloTypes/getTimeseriesData'
+import { getTimeseriesData_getTimeseriesData_items as GetTimeseriesDataType } from '@app/modules/charts/handlers/apolloTypes/getTimeseriesData'
 import { getTimeseriesData, timeseriesEthAvg } from '@app/modules/charts/handlers/timeseriesData.graphql'
 import { Footnote } from '@app/core/components/props'
 
@@ -62,7 +64,7 @@ const DEFAULT_DATA: ComponentDataInterface = {
         start: 24,
         max_items: 24,
         timeOptions: {
-            unit: 'minute',
+            unit: 'hour',
             displayFormats: {
                 minute: 'h:mm a'
             },
@@ -75,7 +77,7 @@ const DEFAULT_DATA: ComponentDataInterface = {
         start: 168,
         max_items: 168,
         timeOptions: {
-            unit: 'day',
+            unit: 'hour',
             displayFormats: {
                 day: 'ddd, h:mm a'
             },
@@ -139,6 +141,9 @@ const DEFAULT_DATA: ComponentDataInterface = {
             variables() {
                 return this.getQueryVars(this.chartKey, this.scale, this.start)
             },
+            skip() {
+                return this.loadingOffset
+            },
             subscribeToMore: {
                 document: timeseriesEthAvg,
                 variables() {
@@ -152,8 +157,8 @@ const DEFAULT_DATA: ComponentDataInterface = {
                             return previousResult
                         }
                         const newItems = this.addSubscriptionItem(newItem, prevItems)
-                        if (newItems.length > this.maxItems) {
-                            newItems.slice(1)
+                        while (newItems.length > this.maxItems) {
+                            newItems.shift()
                         }
                         return {
                             getTimeseriesData: {
@@ -167,13 +172,19 @@ const DEFAULT_DATA: ComponentDataInterface = {
             },
             update: data => (data.getTimeseriesData ? data.getTimeseriesData.items : null),
             result({ data }) {
-                this.chartDataSet = [...this.mapItemsToDataSet(data.getTimeseriesData.items, this.value_type)]
+                if (data && data.getTimeseriesData && data.getTimeseriesData.items && data.getTimeseriesData.items.length > 0) {
+                    this.chartDataSet = [...this.mapItemsToDataSet(data.getTimeseriesData.items, this.value_type)]
+                    this.loadingData = false
+                }
+            },
+            error(error) {
+                this.hasError = true
             }
         }
     }
 })
 export default class TimeSeriesChartData extends Mixins(ChartDataMixin) {
-    @Prop({ type: String, required: true }) chartKey!: TimeseriesKey
+    @Prop({ type: String, required: true }) chartRef!: string
     /*
     ===================================================================================
      Initial Data
@@ -186,6 +197,9 @@ export default class TimeSeriesChartData extends Mixins(ChartDataMixin) {
     maxItems = DEFAULT_DATA[0].max_items
     start = DEFAULT_DATA[0].start
     timeOptions = DEFAULT_DATA[0].timeOptions
+    loadingData = true
+    currentUnit = DEFAULT_DATA[0].timeOptions.unit
+    hasError = false
 
     /*
     ===================================================================================
@@ -193,16 +207,15 @@ export default class TimeSeriesChartData extends Mixins(ChartDataMixin) {
     ===================================================================================
     */
 
-    get chartID(): string {
-        const _id = `${this.chartKey}`
-        return _id.replace(/_/g, '-').toLowerCase()
+    get chartKey(): TimeseriesKey {
+        return ChartRouteKey[this.chartRef].key
     }
 
     get title(): string {
-        return `${this.$t(`charts.${this.chartID}.title`)}`
+        return `${this.$t(`charts.${this.chartRef}.title`)}`
     }
     get description(): string {
-        return `${this.$t(`charts.${this.chartID}.description`)}`
+        return `${this.$t(`charts.${this.chartRef}.description`)}`
     }
     get value_type(): TimeseriesValue {
         return TimeseriesValueKey[this.chartKey]
@@ -216,7 +229,7 @@ export default class TimeSeriesChartData extends Mixins(ChartDataMixin) {
         if (this.chartDataSet) {
             _datasets.push({
                 data: this.chartDataSet,
-                label: `${this.$t(`charts.${this.chartID}.label`)}`,
+                label: `${this.$t(`charts.${this.chartRef}.label`)}`,
                 backgroundColor: 'rgba(118, 221, 251, 0.2)',
                 borderColor: '#2c82be',
                 borderWidth: 2
@@ -257,7 +270,7 @@ export default class TimeSeriesChartData extends Mixins(ChartDataMixin) {
                         display: true,
                         scaleLabel: {
                             display: true,
-                            labelString: `${this.$t(`charts.${this.chartID}.label`)}`
+                            labelString: `${this.$t(`charts.${this.chartRef}.label`)}`
                         },
                         ticks: {
                             source: 'auto',
@@ -270,6 +283,9 @@ export default class TimeSeriesChartData extends Mixins(ChartDataMixin) {
             }
         }
     }
+    get loading(): boolean {
+        return this.loadingOffset || this.loadingData || this.hasError
+    }
 
     /*
     ===================================================================================
@@ -281,11 +297,14 @@ export default class TimeSeriesChartData extends Mixins(ChartDataMixin) {
      * @param value {Number}
      */
     setTimeFrame(value: number): void {
+        if (this.currentUnit !== DEFAULT_DATA[value].timeOptions.unit) {
+            this.loadingData = true
+            this.currentUnit = DEFAULT_DATA[value].timeOptions.unit
+        }
         this.scale = DEFAULT_DATA[value].scale
         this.maxItems = DEFAULT_DATA[value].max_items
         this.start = DEFAULT_DATA[value].start
         this.timeOptions = DEFAULT_DATA[value].timeOptions
-        this.$apollo.queries.getChartData.refetch()
     }
 }
 </script>